dashboard/requirements.txt
redshift-connector
pandas
altair
python-dotenv
streamlit